+++ fuel-merkle/src/sparse/merkle_tree.rs
@@ -46,7 +46,11 @@ impl<TableType, StorageType> MerkleTree<TableType, StorageType> {
     }
 
     pub fn root(&self) -> Bytes32 {
-        self.root_node().hash()
+        *self.root_node().hash()
+    }
+
+    pub fn into_storage(self) -> StorageType {
+        self.storage
     }
 
     // PRIVATE
@@ -122,6 +126,13 @@ where
     TableType: Mappable<Key = Bytes32, Value = Primitive, OwnedValue = Primitive>,
     StorageType: StorageMutate<TableType, Error = StorageError>,
 {
+    /// Build a sparse Merkle tree from a set of key-value pairs. This is
+    /// equivalent to creating an empty sparse Merkle tree and sequentially
+    /// calling [update](Self::update) for each key-value pair. This constructor
+    /// is more performant than calling individual sequential updates and is the
+    /// preferred approach when the key-values are known upfront. Leaves can be
+    /// appended to the returned tree using `update` to further accumulate leaf
+    /// data.
     pub fn from_set<I, D>(mut storage: StorageType, set: I) -> Result<Self, StorageError>
     where
         I: Iterator<Item = (Bytes32, D)>,
@@ -137,7 +148,7 @@ where
 
         for branch in branches.iter() {
             let leaf = &branch.node;
-            storage.insert(&leaf.hash(), &leaf.as_ref().into())?;
+            storage.insert(leaf.hash(), &leaf.as_ref().into())?;
         }
 
         if branches.is_empty() {
@@ -239,7 +250,7 @@ where
         let placeholders = iter::repeat(Node::create_placeholder()).take(depth);
         for placeholder in placeholders {
             node = Node::create_node_on_path(&path, &node, &placeholder);
-            storage.insert(&node.hash(), &node.as_ref().into())?;
+            storage.insert(node.hash(), &node.as_ref().into())?;
         }
 
         let tree = Self {
@@ -259,7 +270,7 @@ where
         }
 
         let leaf_node = Node::create_leaf(key, data);
-        self.storage.insert(&leaf_node.hash(), &leaf_node.as_ref().into())?;
+        self.storage.insert(leaf_node.hash(), &leaf_node.as_ref().into())?;
 
         if self.root_node().is_placeholder() {
             self.set_root_node(leaf_node);
@@ -329,7 +340,7 @@ where
             if !actual_leaf_node.is_placeholder() {
                 current_node = Node::create_node_on_path(path, &current_node, actual_leaf_node);
                 self.storage
-                    .insert(&current_node.hash(), &current_node.as_ref().into())?;
+                    .insert(current_node.hash(), &current_node.as_ref().into())?;
             }
 
             // Merge placeholders
@@ -340,7 +351,7 @@ where
             for placeholder in placeholders {
                 current_node = Node::create_node_on_path(path, &current_node, &placeholder);
                 self.storage
-                    .insert(&current_node.hash(), &current_node.as_ref().into())?;
+                    .insert(current_node.hash(), &current_node.as_ref().into())?;
             }
         }
 
@@ -348,7 +359,7 @@ where
         for side_node in side_nodes {
             current_node = Node::create_node_on_path(path, &current_node, side_node);
             self.storage
-                .insert(&current_node.hash(), &current_node.as_ref().into())?;
+                .insert(current_node.hash(), &current_node.as_ref().into())?;
         }
 
         self.set_root_node(current_node);
@@ -363,7 +374,7 @@ where
         side_nodes: &[Node],
     ) -> Result<(), StorageError> {
         for node in path_nodes {
-            self.storage.remove(&node.hash())?;
+            self.storage.remove(node.hash())?;
         }
 
         let path = requested_leaf_key;
@@ -402,7 +413,7 @@ where
                 if let Some(side_node) = side_nodes_iter.find(|side_node| !side_node.is_placeholder()) {
                     current_node = Node::create_node_on_path(path, &current_node, side_node);
                     self.storage
-                        .insert(&current_node.hash(), &current_node.as_ref().into())?;
+                        .insert(current_node.hash(), &current_node.as_ref().into())?;
                 }
             }
         }
@@ -411,7 +422,7 @@ where
         for side_node in side_nodes_iter {
             current_node = Node::create_node_on_path(path, &current_node, side_node);
             self.storage
-                .insert(&current_node.hash(), &current_node.as_ref().into())?;
+                .insert(current_node.hash(), &current_node.as_ref().into())?;
         }
 
         self.set_root_node(current_node);
@@ -798,16 +809,16 @@ mod test {
 
         let leaf_2_key = leaf_1.hash();
         let leaf_2_data = b"DATA_2";
-        let leaf_2 = Node::create_leaf(&leaf_2_key, leaf_2_data);
+        let leaf_2 = Node::create_leaf(leaf_2_key, leaf_2_data);
 
-        tree.update(&leaf_2_key, leaf_2_data).unwrap();
+        tree.update(leaf_2_key, leaf_2_data).unwrap();
         tree.update(&leaf_1_key, leaf_1_data).unwrap();
         assert_eq!(
-            tree.storage.get(&leaf_2.hash()).unwrap().unwrap().into_owned(),
+            tree.storage.get(leaf_2.hash()).unwrap().unwrap().into_owned(),
             leaf_2.as_ref().into()
         );
         assert_eq!(
-            tree.storage.get(&leaf_1.hash()).unwrap().unwrap().into_owned(),
+            tree.storage.get(leaf_1.hash()).unwrap().unwrap().into_owned(),
             leaf_1.as_ref().into()
         );
     }